+++ packages/blocks-core/js/libs/wordpress/code/index.js
@@ -9,10 +9,19 @@ import { SharedBlockExtension } from '@blockera/editor';
  * Internal dependencies
  */
 import type { BlockType } from '../../../type';
+import sharedInnerBlocks from '../inners/shared';
 
 export const Code: BlockType = {
 	name: 'blockeraCode',
 	targetBlock: 'core/code',
+	blockeraInnerBlocks: {
+		'elements/bold': sharedInnerBlocks['elements/bold'],
+		'elements/italic': sharedInnerBlocks['elements/italic'],
+		'elements/kbd': sharedInnerBlocks['elements/kbd'],
+		'elements/code': sharedInnerBlocks['elements/code'],
+		'elements/span': sharedInnerBlocks['elements/span'],
+		'elements/mark': sharedInnerBlocks['elements/mark'],
+	},
 	edit: (props) => {
 		return <SharedBlockExtension {...props} />;
 	},