+++ packages/blocks-core/CHANGELOG.md
@@ -49,6 +49,7 @@
 - `Term Description` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 - `Term Template` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 - `Verse` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
+- `Code` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 
 
 ### Automated Tests
@@ -96,6 +97,7 @@
 - Added E2E tests to check functionality `Term Description` new inner block.
 - Added E2E tests to check functionality `Term Template` new inner block.
 - Added E2E tests to check functionality `Verse` new inner block.
+- Added E2E tests to check functionality `Code` new inner block.
 
 ## 1.4.0 (2025-05-19)